pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "lxml>=6.0.2",
    "pillow>=12.0.0",
    "python-docx>=1.2.0",
]
